pyproject.toml
package = false

[project]
name="rules_tap"
version="0.0.1"
requires-python = ">=3.13"
dependencies = [
    "colorama==0.4.6",
    "django==5.2.0",
    "faiss-cpu==1.10.0",
    "gunicorn==23.0.0",
    "inflection==0.5.1",
    "openai==1.75.0",
    "psycopg2-binary==2.9.10",
]
